ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
transformers==4.36.2
torch==2.1.2
pillow==10.2.0
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
